owut: disable version_code in build request
Don't add 'version_code' to the build request json.  Seems to work
just fine without it, and it causes the ASU builder fail with the
below error on occasion.  When it's not present, then the build
goes through and installing the resulting image has worked without
issue (only tested twice, both with success).

  Detail: Error: Received inncorrect version r26171-3f28c422ba (requested r26162-6bb2551343)

For reference, the Firmware Selector does not include this property,
but 'auc' does add it and then reports:

  WARNING: cannot determing currently running branch.
  Invalid argument (22)

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>