owut: fix 'what_depends' checking
We were missing some dependencies, so that the top-level list included some
packages that were actually not top-level.  This added them to the build list,
which didn't cause any issue (the would be installed anyway by those exact
dependencies), but it was incorrect.  Also, when doing '--remove' on a
depending package, the missed dependency would not be removed.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>