owut: use downloads site as source for platform data
The ASU server site sometimes lags behind the downloads site by
days, and the metadata for the device is out-of-date with respect
to the imagebuilder.  Specifically, the 'version_code' value is
wrong, which results in errors like the following, even though
everything else matches (and the build would proceed just fine,
if not for this single mismatch).

  Error: Received incorrect version r26474-87d5d38e2a (requested r26467-50e26e6fd7)

It is now switchable, with the 'use_asu' boolean in the 'collect_all'
function, which will probably be removed sometime soon, leaving
just the downloads variant.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>