owut: maintain internal package list in installation order
This addresses a bug in opkg, where if you install certain packages the order
in which you install them changes the outcome (often causing a failure to
install a desired package).  For example, see
https://forum.openwrt.org/t/owut-openwrt-upgrade-tool/200035/61
where the avahi-* packages are the culprit.  There are other reports like this,
and the workaround is to retain the original installation order.

The user reports are still sorted for ease of reading, but the internal lists
generated for the ASU builder retain the installation order.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>