owut: allow user to specify 'version_code' for build request
First mitigation in this commit is to use upstream servers to collect
'version_code', this avoids any lag that the ASU server has in updating its
version data.  This should resolve a lot of issues.

If that fails, allow the user to specify the 'version_code' that is sent to
the ASU server.  A special value of 'none' is treated as meaning "send no
version_code value."

It is often the case that users don't care what version they are getting as long
as it's relatively recent, and the lack of a download simply frustrates them.
'owut' still does a final version code check, but *after* the download, this
gives users the choice to decide if they want to install the "bad" firmware.bin
or ignore it.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>