owut: translate mapped names from platform 'supported_devices'
Certain devices report a descriptive alias in 'board_name', and must be
translated to the actual profile name.  For example, 'ubus call system board'
reports the alias 'raspberrypi,model-b-plus', but the true platform name
is 'rpi'.  Add code to do the translation, using data found in the target
platform.json.

Reported-by: @edrikk
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>