owut: prefer use of packages_changes when generating build request
With the use of 'request.packages', the build signature on the
ASU server doesn't change even when a package is rebuilt, instead
the user (occasionally) receives their previously built image
from the ASU cache due to this unchanged signature.  By using
'request.packages_versions' instead, we generate a different build
signature and avoid the cache.

Note that use of a complete 'packages_signatures' is not always
possible, as historically neither 'kernel' nor 'libc' were included
in the package index.

Fixes: https://github.com/efahl/owut/issues/31
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>