owut: create uclient singleton to minimize memory usage
As of f049043e, every call to uclient.init_ssl has increased
the process's RSS by about 900KB, so if a build requires 100 api
calls to update the status monitor, it would consume 90MB of device
physical memory.  This is causing many users to see "Killed..." when
attempting to run 'owut upgrade'.

Fix this by creating a single uclient object with a single SSL
context, which is reused for very request.  This drops total RSS
from 100+MB down to below 10MB for typical builds.

Link: https://forum.openwrt.org/t/owut-openwrt-upgrade-tool/200035/871
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>